refactor(routing): split routes into individual pages
