+++ src/pages/hello/components/Hello.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from '../../components/Text';
+import { Text } from '../../../components/Text';
 
 interface Props {}
