+++ src/pages/login/LoginRoutes.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Route } from '../../routing/routing-type';
+import { Login } from './components/Login';
+
+export const loginRoutes: Route[] = [
+  {
+    path: '/counter',
+    action(context) {
+      return {
+        content: <Login />,
+      };
+    },
+  },
+];
